refactor(classic): add dp and greedy in classic
